% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_api.R
\name{ppc_api_key}
\alias{ppc_api_key}
\title{Set ProPublica API key}
\usage{
ppc_api_key(api_key = NULL, set_renv = FALSE)
}
\arguments{
\item{api_key}{The actual API key string provided by ProPublica..}

\item{set_renv}{Logical indicating whether to append the environment variable
to the home directory's ".Renviron" file. This requires directory
permission, but it will ensure that future sessions automatically find the
token. The default is FALSE (do NOT set in ~/.Renviron).}
}
\value{
the API key string
}
\description{
Stores ProPublica API key as an environment variable for current and/or
future sessions
}
\examples{

## this is not a real key
ppc_api_key("as9d78f6aayd9fy2fq378a9ds876fsas89d7f")

}
