% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_parse.R
\name{ppc_parse_data}
\alias{ppc_parse_data}
\title{Parse data}
\usage{
ppc_parse_data(r)
}
\arguments{
\item{r}{Reponse from ProPublica API}
}
\value{
A tibble with headers attribute
}
\description{
Converts ProPublica API response into a tibble (data frame)
}
