% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_statements.R
\name{ppc_statements}
\alias{ppc_statements}
\title{Statements}
\usage{
ppc_statements(date = NULL, api_key = NULL, raw = FALSE)
}
\arguments{
\item{api_key}{The actual API key string provided by ProPublica.}

\item{raw}{Logical indicating whether to return the raw response object. The
default (FALSE) parses the content and returns a tibble data frame.}
}
\description{
Get Recent Congressional Statements
}
\details{
To get lists of recent statements published on congressional websites, use the following URI structure. This request returns the 20 most recent results and supports pagination using multiples of 20.
}
\examples{
\dontrun{
ppc_votes()
}
}
