% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_bills.R
\name{ppc_bills}
\alias{ppc_bills}
\title{Bills}
\usage{
ppc_bills(query = "", api_key = NULL, raw = FALSE)
}
\arguments{
\item{query}{keyword or phrase}

\item{api_key}{The actual API key string provided by ProPublica.}

\item{raw}{Logical indicating whether to return the raw response object. The
default (FALSE) parses the content and returns a tibble data frame.}
}
\value{
A data frame of congressional bills information
}
\description{
Search Bills
}
\examples{
\dontrun{
## get information on legislation about health care
hc <- ppc_bills("health care")
}
}
