% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_parse.R
\name{ppc_parse_data}
\alias{ppc_parse_data}
\title{Parse data}
\usage{
ppc_parse_data(r)
}
\arguments{
\item{r}{Reponse from ProPublica API}
}
\value{
A tibble with headers attribute

A data frame of congressional information
}
\description{
Converts ProPublica API response into a tibble (data frame)
}
\examples{
\dontrun{
## get votes data–return the raw response object
v <- ppc_votes("senate", raw = TRUE)

## convert response object into data frame
ppc_parse_data(v)
}
}
