% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_statements.R
\name{ppc_statements}
\alias{ppc_statements}
\title{Statements}
\usage{
ppc_statements(date = NULL, api_key = NULL, raw = FALSE)
}
\arguments{
\item{date}{Day from which statements should be collected; should follow the
format of YYYY-MM-DD.}

\item{api_key}{The actual API key string provided by ProPublica.}

\item{raw}{Logical indicating whether to return the raw response object. The
default (FALSE) parses the content and returns a tibble data frame.}
}
\value{
A data frame of congressional statements information
}
\description{
Get Recent Congressional Statements
}
\examples{
\dontrun{
## get all statements from may 7th, 2018
may72018 <- ppc_statements("2018-05-07")
}
}
